did first half of brand page. Will create full-brand list filter tomorrow
