+++ app/locations/page.module.css
@@ -50,8 +50,7 @@
 .sectionSubtitle {
   font-size: 1.1rem;
   font-weight: bold;
-  color: #666;
-  margin-bottom: 1.5rem;
+  margin-bottom: 1rem;
 }
 
 .formGroup {
