+++ app/styles/Header.css
@@ -173,7 +173,7 @@
 /* -------------------------- */
 .nav-bar {
   background: white;
-  border-top: 1px solid #ddd;
+  border-bottom: 1px solid #ddd;
   padding: 0;
 }
 
@@ -364,7 +364,7 @@
 }
 @media (min-width: 1400px) {
   .dropdown-menu {
-    min-width: 1280px;
+    min-width: 1290px;
   }
 }
