+++ app/styles/LandingPage.css
@@ -64,7 +64,7 @@
   
   /* Skeleton for Category Header */
   .category-page-header.skeleton {
-    height: 40px;
+    height: 77px;
     margin-bottom: 20px;
   }
